(SerialIface) modularize and add support for vgm and vgz.
