+++ examples/SerialIface/exc.py
@@ -1,0 +1,2 @@
+class InvalidFormatError(Exception):
+  pass
