--- examples/SerialIface/play_imf.py
@@ -1,162 +1,0 @@
-# Stream an IMF file via the serial interface of the Arduino OPL shield.
-
-import collections
-import struct
-import sys
-
-import serial
-
-class ArduinoOpl:
-
-  STARTUP_MSG = b'HLO!\n'
-  READY_CMD = b'RDY?\n'
-  READY_RSP = b'RDY!\n'
-  ACK_RSP = b'k'
-  RESET_CMD = b'\x00' * 4
-
-  def __init__(self, portname, baudrate=115200, max_write_ahead=5, debug=False):
-    self.port = serial.Serial(portname, baudrate, timeout=None)
-    self.ready = False
-    self.max_write_ahead = max_write_ahead
-    self.n_outstanding = 0
-    self.debug = debug
-
-    self._status(b'')
-    # Opening port resets Arduino. Wait for ready message.
-    self.wait_for_rsp(self.STARTUP_MSG)
-
-    self._debug('Tx: %s' % self.READY_CMD)
-    self._status(self.READY_CMD)
-    self.port.write(self.READY_CMD)
-    self.wait_for_rsp(self.READY_RSP)
-
-    self.ready = True
-
-  def wait_for_rsp(self, rsp):
-    self._debug('Awaiting: %s' % rsp)
-    rx = self.port.readline()
-    self._debug('Rx: %s' % rx)
-    if not rx.endswith(rsp):
-      raise RuntimeError('Expected: %s, received: %s' % (rsp, rx))
-
-  def write_reg(self, addr, data, delay_ms):
-    if self.n_outstanding >= self.max_write_ahead:
-      rsp = self.port.read()
-      if rsp != self.ACK_RSP:
-        raise RuntimeError('Expected: %s, received: %s' % (rsp, self.ACK_RSP))
-      self.n_outstanding -= 1
-    self.write_reg_unbuffered(addr, data, delay_ms)
-
-  def write_reg_unbuffered(self, addr, data, delay_ms):
-    cmd = struct.pack('BBH', addr, data, delay_ms)
-    self.n_outstanding += 1
-    self.port.write(cmd)
-    self._debug('Tx: %s' % ['%02x' % b for b in cmd])
-    if delay_ms > 0:
-      self._debug('%8d ms' % delay_ms)
-    self._status(cmd)
-
-  def _status(self, last_tx):
-    status = 'Initialized' if self.ready else 'Initializing'
-    buf_status = '[ %s ]' % ('#' * self.n_outstanding).ljust(self.max_write_ahead)
-    tx_txt = 'Tx: %s' % ['%02x' % b for b in last_tx]
-    status_str = "STATUS %-14s BUFFER %-14s %s" % (status, buf_status, tx_txt)
-    print("%s\r" % status_str.ljust(79), end='')
-
-  def _debug(self, txt):
-    if self.debug:
-      print(txt)
-
-  def close(self):
-    self.port.write(self.RESET_CMD)
-    self._status(self.RESET_CMD)
-    self.port.close()
-
-
-IMF_FREQUENCIES = collections.defaultdict(lambda: 560)
-IMF_FREQUENCIES.update({
-    'duke': 280,
-    'wolf': 700
-})
-
-def play_imf(opl, imf_stream, frequency_hz):
-  song_length, = struct.unpack('H', imf_stream.read(2))
-
-  if song_length == 0:
-    imf_type = 0
-    imf_stream.read(2)
-  else:
-    imf_type = 1
-
-  i = 0
-  while imf_type == 0 or imf_type == 1 and i < song_length:
-    cmd = imf_stream.read(4)
-    if len(cmd) < 4:
-      break
-
-    addr, data, delay_cycles = struct.unpack_from('BBH', cmd)
-    delay_us = 1000000 * delay_cycles // frequency_hz
-    opl.write_reg(addr, data, delay_us // 1000)
-    i += 4
-
-
-def usage():
-  print(
-      '''Arduino OPL2 streaming player
-
-Usage:
-   %s <serial_port> [<imf_file> [<imf_frequency>]]
-
-   - serial_port: the name of the Arduino USB serial port
-   - imf_file: the path to an IMF file
-   - imf_frequency:
-     * 'wolf' for Wolfenstein 3d engine music
-     * 'duke' for Duke Nukem music
-     * an integer value to specify frequency in Hertz
-
-Examples:
-
-  %s COM4
-  - Connect to and initialize the Arduino board on COM4
-
-  %s COM4 k5t06.imf
-  - Play k5t06.imf at the default frequency
-
-  %s COM4 dukeiia.imf duke
-  - Play dukeiia.imf at the Duke Nukem frequency
-
-  %s COM4 pow.wlf 700
-  - Play pow.wlf at 700 Hz
-
-''' % ((sys.argv[0],) * 5)
-  )
-  sys.exit()
-
-
-def handle_arguments():
-  portname = sys.argv[1]
-  try:
-    if len(sys.argv) > 2:
-      filename = sys.argv[2]
-
-      frequency_hz = IMF_FREQUENCIES['default']
-      if len(sys.argv) > 3:
-        try:
-          frequency_hz = int(sys.argv[3])
-        except ValueError:
-          frequency_hz = IMF_FREQUENCIES[sys.argv[3].lower()]
-
-      with open(sys.argv[2], 'rb') as f:
-        print('Play "%s" on %s @ %d Hz' % (filename, portname, frequency_hz))
-        print('(Ctrl-C to stop)')
-        opl = ArduinoOpl(portname)
-        play_imf(opl, f, frequency_hz)
-
-  except (KeyboardInterrupt, SystemExit):
-    opl.close()
-
-
-if __name__ == '__main__':
-  if len(sys.argv) <= 1:
-    usage()
-  handle_arguments()
